# artifacts/api-server/src/routes/auth.ts
import { Router, type IRouter } from "express";
import { db, systemUsersTable, activityLogsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import bcrypt from "bcryptjs";
import jwt from "jsonwebtoken";
import { parseBrowser, parseDevice, getClientIp } from "../middlewares/activityLogger";

const router: IRouter = Router();
const JWT_SECRET = process.env.JWT_SECRET || "employee-master-secret-2024";

router.post("/login", async (req, res) => {
  try {
    const { email, password } = req.body;
    if (!email || !password) {
      return res.status(400).json({ error: "Email and password required" });
    }
    const users = await db.select().from(systemUsersTable).where(eq(systemUsersTable.email, email));
    const user = users[0];
    if (!user) {
      return res.status(401).json({ error: "Invalid credentials" });
    }
    const valid = await bcrypt.compare(password, user.passwordHash);
    if (!valid) {
      return res.status(401).json({ error: "Invalid credentials" });
    }
    const token = jwt.sign({ userId: user.id, email: user.email, role: user.role }, JWT_SECRET, { expiresIn: "7d" });

    const ua = req.headers["user-agent"] || "";
    db.insert(activityLogsTable).values({
      userId: user.id,
      userEmail: user.email,
      action: "LOGIN",
      eventType: "User Login",
      entity: "auth",
      entityId: null,
      description: `User login: ${user.email}`,
      browser: parseBrowser(ua),
      device: parseDevice(ua),
      browserIp: getClientIp(req),
    }).catch(console.error);

    return res.json({
      token,
      user: { id: user.id, name: user.name, email: user.email, role: user.role, createdAt: user.createdAt.toISOString() },
    });
  } catch (err) {
    console.error(err);
    return res.status(500).json({ error: "Internal server error" });
  }
});

router.post("/logout", (_req, res) => {
  return res.json({ success: true, message: "Logged out" });
});

router.get("/me", async (req, res) => {
  try {
    const authHeader = req.headers.authorization;
    if (!authHeader?.startsWith("Bearer ")) {
      return res.status(401).json({ error: "Unauthorized" });
    }
    const token = authHeader.slice(7);
    const decoded = jwt.verify(token, JWT_SECRET) as { userId: number };
    const users = await db.select().from(systemUsersTable).where(eq(systemUsersTable.id, decoded.userId));
    const user = users[0];
    if (!user) {
      return res.status(401).json({ error: "User not found" });
    }
    return res.json({ id: user.id, name: user.name, email: user.email, role: user.role, createdAt: user.createdAt.toISOString() });
  } catch {
    return res.status(401).json({ error: "Unauthorized" });
  }
});

export default router;