# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Employee Master API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface SystemUser {
  id: number;
  name: string;
  email: string;
  role: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: SystemUser;
}

export interface CreateSystemUserRequest {
  name: string;
  email: string;
  password: string;
  role: string;
}

export interface UpdateSystemUserRequest {
  name?: string;
  email?: string;
  password?: string;
  role?: string;
}

export interface EmployeeAccess {
  zohoEmail?: boolean;
  microsoftEmail?: boolean;
  microsoftOffice?: boolean;
  finfluxBmDashboard?: boolean;
  mobiliteField?: boolean;
  mobiliteCredit?: boolean;
  hoDashboard?: boolean;
  lightMoney?: boolean;
  zohoProjects?: boolean;
  jira?: boolean;
  bitbucket?: boolean;
  adobeAcrobat?: boolean;
  assetcuez?: boolean;
  exotel?: boolean;
  godaddy?: boolean;
  bluehost?: boolean;
  hostinger?: boolean;
  emailHosting?: boolean;
  awsConsole?: boolean;
  msg91?: boolean;
  dmsAlfresco?: boolean;
}

export interface Employee {
  id: number;
  serialNumber?: number;
  employeeCode: string;
  name: string;
  status: string;
  department?: string;
  designation?: string;
  contactNumber?: string;
  state?: string;
  branch?: string;
  joiningDate?: string;
  email?: string;
  requester?: string;
  approver?: string;
  creator?: string;
  exitInitiator?: string;
  exitDate?: string;
  userExitStatus?: string;
  access?: EmployeeAccess;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEmployeeRequest {
  employeeCode: string;
  name: string;
  status: string;
  department?: string;
  designation?: string;
  contactNumber?: string;
  state?: string;
  branch?: string;
  joiningDate?: string;
  email?: string;
  requester?: string;
  approver?: string;
  creator?: string;
  exitInitiator?: string;
  exitDate?: string;
  userExitStatus?: string;
}

export interface UpdateEmployeeRequest {
  employeeCode?: string;
  name?: string;
  status?: string;
  department?: string;
  designation?: string;
  contactNumber?: string;
  state?: string;
  branch?: string;
  joiningDate?: string;
  email?: string;
  requester?: string;
  approver?: string;
  creator?: string;
  exitInitiator?: string;
  exitDate?: string;
  userExitStatus?: string;
}

export interface UpdateAccessRequest {
  access: EmployeeAccess;
}

export interface ImportEmployeesRequest {
  employees: CreateEmployeeRequest[];
  updateOnDuplicate?: boolean;
}

export type ImportResultErrorsItem = {
  row?: number;
  error?: string;
};

export interface ImportResult {
  imported: number;
  updated: number;
  failed: number;
  errors: ImportResultErrorsItem[];
}

export type DashboardStatsByDepartmentItem = {
  department?: string;
  count?: number;
};

export type DashboardStatsByBranchItem = {
  branch?: string;
  count?: number;
};

export type DashboardStatsAccessDistributionItem = {
  name?: string;
  count?: number;
};

export interface DashboardStats {
  totalEmployees: number;
  activeEmployees: number;
  exitedEmployees: number;
  pendingExitRequests: number;
  accessSystemUsers: number;
  byDepartment: DashboardStatsByDepartmentItem[];
  byBranch: DashboardStatsByBranchItem[];
  accessDistribution: DashboardStatsAccessDistributionItem[];
}

export interface BranchFileStation {
  id: number;
  username: string;
  password?: string;
  groupEmail?: string;
  branchName: string;
  createdAt: string;
}

export interface CreateBranchFileStationRequest {
  username: string;
  password?: string;
  groupEmail?: string;
  branchName: string;
}

export interface AssetCuez {
  id: number;
  employeeId: string;
  firstName: string;
  lastName?: string;
  contactNumber?: string;
  activationStatus?: string;
  createdAt: string;
}

export interface CreateAssetCuezRequest {
  employeeId: string;
  firstName: string;
  lastName?: string;
  contactNumber?: string;
  activationStatus?: string;
}

export interface VpnUser {
  id: number;
  employeeId: string;
  username: string;
  department?: string;
  designation?: string;
  userId?: string;
  password?: string;
  status?: string;
  createdAt: string;
}

export interface CreateVpnRequest {
  employeeId: string;
  username: string;
  department?: string;
  designation?: string;
  userId?: string;
  password?: string;
  status?: string;
}

export interface JiraUser {
  id: number;
  employeeId: string;
  username: string;
  email?: string;
  userStatus?: string;
  addedToOrganization?: string;
  createdAt: string;
}

export interface CreateJiraRequest {
  employeeId: string;
  username: string;
  email?: string;
  userStatus?: string;
  addedToOrganization?: string;
}

export interface MailVaultUser {
  id: number;
  employeeId: string;
  username: string;
  userId?: string;
  userRole?: string;
  password?: string;
  createdAt: string;
}

export interface CreateMailVaultRequest {
  employeeId: string;
  username: string;
  userId?: string;
  userRole?: string;
  password?: string;
}

export interface FtpUser {
  id: number;
  username: string;
  password?: string;
  port?: string;
  hostname?: string;
  path?: string;
  createdAt: string;
}

export interface CreateFtpRequest {
  username: string;
  password?: string;
  port?: string;
  hostname?: string;
  path?: string;
}

export interface AcronisRecord {
  id: number;
  employeeId: string;
  employeeName: string;
  department?: string;
  designation?: string;
  email?: string;
  hostname?: string;
  backupPlan?: string;
  monthlyFullBackupDate?: string;
  backupPlanName?: string;
  backupTime?: string;
  incrementalBackup?: string;
  cpuPriority?: string;
  installationStatus?: string;
  policyAssign?: string;
  backupStatus?: string;
  leftEncryptionPassword?: string;
  createdAt: string;
}

export interface CreateAcronisRequest {
  employeeId: string;
  employeeName: string;
  department?: string;
  designation?: string;
  email?: string;
  hostname?: string;
  backupPlan?: string;
  monthlyFullBackupDate?: string;
  backupPlanName?: string;
  backupTime?: string;
  incrementalBackup?: string;
  cpuPriority?: string;
  installationStatus?: string;
  policyAssign?: string;
  backupStatus?: string;
  leftEncryptionPassword?: string;
}

export interface TataTeleRecord {
  id: number;
  employeeId: string;
  employeeName: string;
  department?: string;
  designation?: string;
  mobileNumber?: string;
  simNumber?: string;
  plan?: string;
  status?: string;
  createdAt: string;
}

export interface CreateTataTeleRequest {
  employeeId: string;
  employeeName: string;
  department?: string;
  designation?: string;
  mobileNumber?: string;
  simNumber?: string;
  plan?: string;
  status?: string;
}

export type ListEmployeesParams = {
  status?: string;
  department?: string;
  branch?: string;
  search?: string;
};