# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Employee Master API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: branch
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/import:
    post:
      operationId: importEmployees
      tags: [employees]
      summary: Import employees from CSV/XLSX
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportEmployeesRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /employees/{id}/access:
    put:
      operationId: updateEmployeeAccess
      tags: [employees]
      summary: Update employee system access checklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccessRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/SystemUser"
      required:
        - token
        - user
    SystemUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    EmployeeAccess:
      type: object
      properties:
        zohoEmail:
          type: boolean
        microsoftEmail:
          type: boolean
        microsoftOffice:
          type: boolean
        finfluxBmDashboard:
          type: boolean
        mobiliteField:
          type: boolean
        mobiliteCredit:
          type: boolean
        hoDashboard:
          type: boolean
        lightMoney:
          type: boolean
        zohoProjects:
          type: boolean
        jira:
          type: boolean
        bitbucket:
          type: boolean
        adobeAcrobat:
          type: boolean
        assetcuez:
          type: boolean
        exotel:
          type: boolean
        godaddy:
          type: boolean
        bluehost:
          type: boolean
        hostinger:
          type: boolean
        emailHosting:
          type: boolean
        awsConsole:
          type: boolean
        msg91:
          type: boolean
        dmsAlfresco:
          type: boolean
    Employee:
      type: object
      properties:
        id:
          type: integer
        serialNumber:
          type: integer
        employeeCode:
          type: string
        name:
          type: string
        status:
          type: string
        department:
          type: string
        designation:
          type: string
        contactNumber:
          type: string
        state:
          type: string
        branch:
          type: string
        joiningDate:
          type: string
        email:
          type: string
        requester:
          type: string
        approver:
          type: string
        creator:
          type: string
        exitInitiator:
          type: string
        exitDate:
          type: string
        userExitStatus:
          type: string
        access:
          $ref: "#/components/schemas/EmployeeAccess"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - employeeCode
        - name
        - status
        - createdAt
        - updatedAt
    CreateEmployeeRequest:
      type: object
      properties:
        employeeCode:
          type: string
        name:
          type: string
        status:
          type: string
        department:
          type: string
        designation:
          type: string
        contactNumber:
          type: string
        state:
          type: string
        branch:
          type: string
        joiningDate:
          type: string
        email:
          type: string
        requester:
          type: string
        approver:
          type: string
        creator:
          type: string
        exitInitiator:
          type: string
        exitDate:
          type: string
        userExitStatus:
          type: string
      required:
        - employeeCode
        - name
        - status
    UpdateEmployeeRequest:
      type: object
      properties:
        employeeCode:
          type: string
        name:
          type: string
        status:
          type: string
        department:
          type: string
        designation:
          type: string
        contactNumber:
          type: string
        state:
          type: string
        branch:
          type: string
        joiningDate:
          type: string
        email:
          type: string
        requester:
          type: string
        approver:
          type: string
        creator:
          type: string
        exitInitiator:
          type: string
        exitDate:
          type: string
        userExitStatus:
          type: string
    UpdateAccessRequest:
      type: object
      properties:
        access:
          $ref: "#/components/schemas/EmployeeAccess"
      required:
        - access
    ImportEmployeesRequest:
      type: object
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/CreateEmployeeRequest"
        updateOnDuplicate:
          type: boolean
      required:
        - employees
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        updated:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              error:
                type: string
      required:
        - imported
        - updated
        - failed
        - errors
...